refactor: Profile api authorization now is working well with new jwt format and roles

+++ services/profile-api/src/Program.cs
@@ -3,6 +3,9 @@ using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.IdentityModel.Tokens;
 using System.Text;
 using Microsoft.AspNetCore.Authorization;
+using System.IdentityModel.Tokens.Jwt;
+
+JwtSecurityTokenHandler.DefaultInboundClaimTypeMap.Clear();
 
 var builder = WebApplication.CreateBuilder(args);
 
@@ -57,9 +60,14 @@ if (app.Environment.IsDevelopment())
 app.UseAuthentication();
 app.UseAuthorization();
 
+string roleClaimType = "http://schemas.microsoft.com/ws/2008/06/identity/claims/role";
+string subClaimType = "http://schemas.xmlsoap.org/ws/2005/05/identity/claims/nameidentifier";
 
-app.MapGet("/profiles", [Authorize] async (ProfileDbContext db) =>
+app.MapGet("/profiles", [Authorize] async (HttpContext httpContext, ProfileDbContext db) =>
 {
+    var role = httpContext.User.FindFirst(roleClaimType)?.Value;
+    if (role == "Banned") return Results.Forbid();
+
     try
     {
         return Results.Ok(await db.Profiles.ToListAsync());
@@ -70,8 +78,11 @@ app.MapGet("/profiles", [Authorize] async (ProfileDbContext db) =>
     }
 });
 
-app.MapGet("/profiles/{id}", [Authorize] async (int id, ProfileDbContext db) =>
+app.MapGet("/profiles/{id}", [Authorize] async (HttpContext httpContext, string id, ProfileDbContext db) =>
 {
+    var role = httpContext.User.FindFirst(roleClaimType)?.Value;
+    if (role == "Banned") return Results.Forbid();
+
     try
     {
         var profile = await db.Profiles.FindAsync(id);
@@ -83,8 +94,20 @@ app.MapGet("/profiles/{id}", [Authorize] async (int id, ProfileDbContext db) =>
     }
 });
 
-app.MapPost("/profiles", [Authorize] async (Profile profile, ProfileDbContext db) =>
+app.MapPost("/profiles", [Authorize] async (HttpContext httpContext, Profile profile, ProfileDbContext db) =>
 {
+    var user = httpContext.User;
+
+    foreach (var claim in user.Claims)
+    {
+        Console.WriteLine($"[PROFILE-API] Claim: {claim.Type} = {claim.Value}");
+    }
+
+    var role = user.FindFirst(roleClaimType)?.Value;
+    Console.WriteLine($"[PROFILE-API] Rol recibido en token JWT: {role}");
+
+    if (role != "Admin") return Results.Forbid();
+
     try
     {
         db.Profiles.Add(profile);
@@ -97,22 +120,31 @@ app.MapPost("/profiles", [Authorize] async (Profile profile, ProfileDbContext db
     }
 });
 
-app.MapPut("/profiles/{id}", [Authorize] async (int id, Profile updatedProfile, ProfileDbContext db) =>
+app.MapPut("/profiles/{id}", [Authorize] async (HttpContext httpContext, string id, Profile updatedProfile, ProfileDbContext db) =>
 {
+    var user = httpContext.User;
+    var role = user.FindFirst(roleClaimType)?.Value;
+    var sub = user.FindFirst(subClaimType)?.Value;
+
+    if (role is null || sub is null)
+        return Results.Forbid();
+
     try
     {
         var profile = await db.Profiles.FindAsync(id);
         if (profile is null) return Results.NotFound();
 
-        profile.Name = updatedProfile.Name;
-        profile.Photo = updatedProfile.Photo;
-        profile.Email = updatedProfile.Email;
-        profile.PasswordHash = updatedProfile.PasswordHash;
-        profile.Sex = updatedProfile.Sex;
-        profile.Nickname = updatedProfile.Nickname;
+        if (role == "Admin" || (role == "User" && id == sub))
+        {
+            profile.Name = updatedProfile.Name;
+            profile.Photo = updatedProfile.Photo;
+            profile.Sex = updatedProfile.Sex;
 
-        await db.SaveChangesAsync();
-        return Results.NoContent();
+            await db.SaveChangesAsync();
+            return Results.NoContent();
+        }
+
+        return Results.Forbid();
     }
     catch (Exception ex)
     {
@@ -120,16 +152,28 @@ app.MapPut("/profiles/{id}", [Authorize] async (int id, Profile updatedProfile, 
     }
 });
 
-app.MapDelete("/profiles/{id}", [Authorize] async (int id, ProfileDbContext db) =>
+app.MapDelete("/profiles/{id}", [Authorize] async (HttpContext httpContext, string id, ProfileDbContext db) =>
 {
+    var user = httpContext.User;
+    var role = user.FindFirst(roleClaimType)?.Value;
+    var sub = user.FindFirst(subClaimType)?.Value;
+
+    if (role is null || sub is null)
+        return Results.Forbid();
+
     try
     {
         var profile = await db.Profiles.FindAsync(id);
         if (profile is null) return Results.NotFound();
 
-        db.Profiles.Remove(profile);
-        await db.SaveChangesAsync();
-        return Results.NoContent();
+        if (role == "Admin" || (role == "User" && id == sub))
+        {
+            db.Profiles.Remove(profile);
+            await db.SaveChangesAsync();
+            return Results.NoContent();
+        }
+
+        return Results.Forbid();
     }
     catch (Exception ex)
     {
